% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PollutionData.R
\docType{class}
\name{PollutionData-class}
\alias{PollutionData-class}
\alias{PollutionData}
\title{Pollution data}
\description{
Pollution data
}
\section{Fields}{

\describe{
\item{\code{response_list}}{list.}

\item{\code{url}}{character.}

\item{\code{countries}}{list.}
}}


\examples{
PollutionData
}
