% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinnyapp.R
\docType{class}
\name{shinnyapp-class}
\alias{shinnyapp-class}
\alias{shinnyapp}
\title{Title shinny application for world population data}
\description{
Title shinny application for world population data
}
\section{Fields}{

\describe{
\item{\code{server_components}}{list.}
}}


